Serialization/Deserialization done
Serializatio/Deserialization done, messageHandler also done as much can be done without the commandHandler.

+++ Common/Message/Modbus/Modbus PDU/Request/Write/Interface/IModbusWriteMultipleCoilsRequest.cs
@@ -10,7 +10,7 @@ namespace Common.Message
     {
         short StartingAddress { get; set; }
         short QuantityOfOutputs { get; set; }
-        short ByteCount { get; set; }
-        short[] OutputsValue { get; set; }
+        byte ByteCount { get; set; }
+        byte[] OutputsValue { get; set; }
     }
 }
